Move MySqlCompatibleH2DatabaseConfigurer to JdbcUsersConnectionRepositoryTest
